improve devshell caching
Unfortunately it is impossible to guarantee that required flake inputs
are included in the profile closure until we have a solution for
https://github.com/NixOS/nix/issues/3913

Until that time, we can build the entire devshell package and add it as
a root as this ensures more of the dependencies of the build are kept
in the profile gc root.

Also add `--keep-outputs` for the same reason.