properly deprecate the organelle type
Previously, if a user still had any references to an old revision of
standard in their lock file, the types defined using that standard
would fail to evaluate due to the recent rename. This allows the old
type to properly evaluate with yants while informing the user that
downstream may need to update their flake.