feat: allow implicit block API
Cell Blocks always have the same API (i.e. `{inputs, cell}:`).

With this addition any cell block can simply assume that `inputs` &
`cell` are in scope without having to explicitly declare them, removing
the little bit of boilerplate that we did have.

For backwards compatibility, the old API and its validation is
maintained if the file contains a lambda.