feat(devshells): add a devshell drv generator
Currently `nix develop` builds its own ad hoc derivation directly in
source code, meaning this derivation cannot be referenced in Nix code,
and therefore cannot be captured as a proper dependency of any script
that may wish to enter it before execution.

We, therefore, provide a derivation wrapper function that attempts to
rebuild this same exact derivation in proper Nix, making it a regular
referencable value like any other, but with the context of a shell
derivation.

Currently the hashes of the real shell derivation produced by
`nix develop` and this simulated version do not match, only because
the name and outpath do not match, and that is something that is not
duplicable in regular Nix code unfortanately, but a quirck of the C++
implementation.