feat(proviso): implement solid API
A proviso is now a string that defines a bash function "proviso", whose
first & second argument are a nameref to the JSON input and output
lists, respectively.

Users now have more freedom to process the lists however they choose
and simply append the final results of the check to the output list.

This opens the possibility of processing all the check logic in
parallel, for example.