feat(ci): implement proviso for OCI images
Skopeo takes a few seconds to query a single image so in order to avoid
adding significant overhead to discovery, images are checked in
parallel using xargs.

This adds some complexity to the implementation but allows us to check
all images in just a few seconds. That is, if _t_ is the time it takes
to check one image and _n_ is the number of images, this implementation
allows us to pay _t_ instead of _nt_.