fix: mkOperableScript needs args list
An atribute set makes single flag arguments awkward to declare, and
no specific order is guaranteed, meaning that programs that require
args in a particular order will fail.

In order to use the style which is most natural for the given program
and to keep the resulting script readable, users can group key=value
style arguments in a true attribute set, or several arguments of
the same class in a list; i.e:
```
mkOperableScript {
  args = [
    "--single-flag"
    {"--port" = 80;}
    ["path/a" "path/b" "path/c" ["path/with" "spaces"]]
  ];
  # ...
}
```
results in:
```
args+=("--single-flag")
args+=("--port" "80")
args+=("path/a" "path/b" "path/c" "path/with spaces")
```

For the sake of backward compatibility, the previous API is retained
in the event that an attribute set is passed instead of a list.